02-workflows/pyproject.toml
[project]
name = "02-workflows"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
